Simplify feature module registry test
