+++ clang-tools-extra/clangd/unittests/CMakeLists.txt
@@ -155,8 +155,6 @@ target_include_directories(ClangdTests PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}>
 )
 
-add_subdirectory(feature-modules)
-
 clang_target_link_libraries(ClangdTests
   PRIVATE
   clangAST
@@ -186,8 +184,6 @@ target_link_libraries(ClangdTests
   clangTidy
   clangTidyUtils
   clangdSupport
-
-  ${FEATURE_MODULES}
   )
 
 if (CLANGD_ENABLE_REMOTE)
