+++ clang-tools-extra/clangd/unittests/FeatureModulesRegistryTests.cpp
@@ -7,8 +7,8 @@
 //===----------------------------------------------------------------------===//
 
 #include "FeatureModule.h"
-#include "feature-modules/ForceLinker.h" // IWYU pragma: keep
 #include "refactor/Tweak.h"
+#include "support/Logger.h"
 
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
@@ -49,9 +49,35 @@ raw_ostream &operator<<(raw_ostream &OS, const clang::clangd::Tweak &T) {
 namespace clang::clangd {
 namespace {
 
+class Dummy final : public FeatureModule {
+  static constexpr const char *TweakID = "DummyTweak";
+  struct DummyTweak final : public Tweak {
+    const char *id() const override { return TweakID; }
+    bool prepare(const Selection &) override { return true; }
+    Expected<Effect> apply(const Selection &) override {
+      return error("not implemented");
+    }
+    std::string title() const override { return id(); }
+    llvm::StringLiteral kind() const override {
+      return llvm::StringLiteral("");
+    };
+  };
+
+  void contributeTweaks(std::vector<std::unique_ptr<Tweak>> &Out) override {
+    Out.emplace_back(new DummyTweak);
+  }
+};
+
+static FeatureModuleRegistry::Add<Dummy>
+    X("dummy", "Dummy feature module with dummy tweak");
+
 MATCHER_P(moduleName, Name, "") { return arg.getName() == Name; }
 MATCHER_P(tweakID, ID, "") { return arg->id() == llvm::StringRef(ID); }
 
+// In this test, it is assumed that for unittests executable, all feature
+// modules are added to the registry only here (in this file). To implement
+// modules for clangd tool, one need to link them directly to the clangd
+// executable in clangd/tool/CMakeLists.txt.
 TEST(FeatureModulesRegistryTest, DummyModule) {
   EXPECT_THAT(FeatureModuleRegistry::entries(),
               ElementsAre(moduleName("dummy")));
