--- clang-tools-extra/clangd/unittests/feature-modules/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_subdirectory(dummy)
-set(FEATURE_MODULES ${FEATURE_MODULES} PARENT_SCOPE)
